Rebalancing gotcha boss...
I am playing on the easy difficulty, having set every favorable Obsidian option in my favor and every disadvantageous option also to my favor (including minimum boss health, maximum pickups, zero reinforcements...) and it still takes me a `idkfa` and all of my energy ammo to down the boss.

The best weapon the game will give you is a rocket launcher, so the idea that this is beatable with a pistol start seems laughable. Most players are unlikely to finish a full-episode permadeth run with full energy ammo but I guess in that case it should be feasible to finish the boss with rockets after having weakened him.

This is the furthest the options will help make a boss easier. Ammo distribution and overall balance is feeling pretty good for progress throughout the episode so I don't want to make it higher just to guarantee a final boss kill. If this proves to not be reliably beatable outside of testing, disabling the boss level might be necessary in favor of a higher but slow-curve map size range or similar alternative.

At this point I wouldn't even care if beating the boss is just a formality to end the WAD on a high note of spectacle. If the fight can't be properly balanced or at least winnable with a generous amount of pistol restarts, I'd be happy to just make it even easier and let the player enjoy a semi-guaranteed victory.