Reducing enemy count...
Last one was a bit too long due to plenty of dangling enemies coming up one by one. This one feels right length-wise but too easy overall (final boss too) despite having a couple of moments of low-ammo (only early on) and a near-death or two.

Overall, probably an improvement with some tuning up to do. May be a challenge getting to both shorten the WADs by 20% and also increase the difficulty slightly.