Tune difficulty slightly up...
Bit less ammunition and weapon tiers a bit less early. Weapons could be delayed another step but let's test it first as balance is pretty sweet as-is.

I wouldn't change these necessarily but I think they won't make the game deadlier as much as it'll make players think and adapt more.

Overall, I'd say that my completion-rate on normal-difficulty, iron-man is about 50% which is as good as I could hope for but it does feel a bit easy (until you die suddenly). Hopefully this improves game-play without making things strictly harder, so worth trying.

Me getting better at the game could be a variable too but playing on normal should give ample breadth for players to tune it up or down.